11/09/2023

+++ src/UniversidadEjemplo/Entidades/Inscripcion.java
@@ -6,15 +6,15 @@ public class Inscripcion {
     private int idInscripcion;
     private Materia materia;
     private Alumno alumno;
-    private double nota;
+    private int nota;
 
-    public Inscripcion(Materia materia, Alumno alumno, double nota) {
+    public Inscripcion(Materia materia, Alumno alumno, int nota) {
         this.materia = materia;
         this.alumno = alumno;
         this.nota = nota;
     }
 
-    public Inscripcion(int idInscripcion, Materia materia, Alumno alumno, double nota) {
+    public Inscripcion(int idInscripcion, Materia materia, Alumno alumno, int nota) {
         this.idInscripcion = idInscripcion;
         this.materia = materia;
         this.alumno = alumno;
@@ -48,11 +48,11 @@ public class Inscripcion {
         this.alumno = alumno;
     }
 
-    public double getNota() {
+    public int getNota() {
         return nota;
     }
 
-    public void setNota(double nota) {
+    public void setNota(int nota) {
         this.nota = nota;
     }
